api/azure_functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
numpy==1.16.0
tensorflow==1.15.4
keras==2.1.6
h5py~=2.10.0